+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # FEniCS-preCICE adapter changelog
 
+## 1.1.0
+
+* Only warn during initialization, if duplicate boundary point is found for point sources.
+* Remove deprecated package `fenicsadapter`. Don't use `import fenicsadapter`. Please use `import fenicsprecice`. https://github.com/precice/fenics-adapter/pull/121
+
 ## 1.0.1
 
 * Bugfix for PointSources https://github.com/precice/fenics-adapter/issues/109